:art: move jsonld model definitions to own module
